new endpoint for StopSearch added

+++ transit-routing-backend/src/main/java/com/dnikitin/transit/infrastructure/repository/StopJpaRepository.java
@@ -3,6 +3,6 @@ package com.dnikitin.transit.infrastructure.repository;
 import com.dnikitin.transit.infrastructure.persistence.entity.StopEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface StopJpaRepository extends JpaRepository<StopEntity, Integer> {
+public interface StopJpaRepository extends JpaRepository<StopEntity, Long> {
 }
